django-server/requirements.txt
django 
djangorestframework 
django-cors-headers 
djangorestframework-simplejwt
 